requirements-lock.txt
# ToM-NAS Production Dependencies - Pinned Versions
# Generated for reproducible builds and Google Cloud deployment
# Last updated: 2025-12-12

# Core ML Framework
torch==2.1.2
numpy==1.26.2
scipy==1.11.4

# Graph Operations
networkx==3.2.1

# Visualization
matplotlib==3.8.2
plotly==5.18.0
streamlit==1.29.0

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

# Progress & Utilities
tqdm==4.66.1

# Google Cloud Integration
google-cloud-logging==3.8.0
google-cloud-monitoring==2.18.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.17.0

# Production Server
gunicorn==21.2.0
uvicorn==0.25.0

# API Framework (for REST endpoints)
fastapi==0.108.0
pydantic==2.5.3

# Health Checks & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-gcp-trace==1.6.0

# Security
python-dotenv==1.0.0
cryptography==41.0.7

# Testing (dev)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2

# Code Quality (dev)
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
